From 0aaa764b1e3fd7d8816678e3b960199e5595b3eb Mon Sep 17 00:00:00 2001 From: Igor Druzhinin <igor.druzhinin@citrix.com> Date: Fri, 7 Apr 2017 20:31:50 +0100 Subject: [PATCH] CP-21767 Don't accidently unplug ourselves if  PCI_CLASS_DEVICE is 0x0100
Due to some wild legacy reasons we have to set 0x0100 as a PCI class ID
for Xen Platform device which is a PCI class ID of SCSI Storage
Controller. During Linux kernel initialization all the PV and emulated
SCSI controllers are unplugged which leads to the accidental unplug of
Xen Platform device itself.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>