From bad71b32564565d340497a4e29b10577b0239b42 Mon Sep 17 00:00:00 2001 From: Igor Druzhinin <igor.druzhinin@citrix.com> Date: Tue, 20 Jun 2017 17:24:43 +0100 Subject: [PATCH] CA-256542 Workaround unassigned accesses caused by ballooning
Xen Balloon memory hotplug makes a guest to populate memory areas
which are not sufficiently described in QEMU. Some legacy Linux
kernels, in turn, may try to allocate DMA buffers for emulated devices
in these areas. When QEMU receives an address which is not present
in its memory map it marks that as 'unassigned' and avoids any
DMA processing. That eventually results in broken emulation and
guest error reporting.

Since we want to support legacy guests as we did on QEMU-trad we
introduce "allow-unassigned" mode in a form of machine option.
We try to intercept and correctly process all the unassigned memory
accesses in QEMU in this mode.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>