xen: Remove wrapper functions
QEMU uses pkg-config to detect the Xen version and then uses that to
determine which library functions are present and which need wrappers.
This doesn't work with our Xen due to the presence of certain backports,
so remove wrapper functions that are not needed because they've been
backported to our Xen.